util/bg: do not check boundary characters
Some languages do not use space character to separate words, so
when grep keep words (ascii characters) from the translated message,
we need to make sure the extracted message is a proper segment of the
original message. E.g. in vi (Vietnamese) translation, has an entry
like below:

    msgid: create/reset and checkout a branch
    msgstr: tạo/đặt_lại và checkout một nhánh"

When searching keep words in msgstr, we may get a variable like
keep word "t_l", but by checking boundary of the word (đặt_lại),
we can exclude this false positive matching result.

But for bg (Bulgarian) translations, the "<>" around the placeholders
are removed, we should turn off this check for bg translations.

Signed-off-by: Jiang Xin <zhiyou.jx@alibaba-inc.com>